fix: nested list background color overflow
